requirements.txt
Flask
gunicorn
requests
google-cloud-speech